requirements.txt
fastapi
uvicorn
python-docx
openai
python-multipart
python-dotenv
